+++ src/cmdorc/__init__.py
@@ -1,6 +1,6 @@
 __version__ = "0.2.1"
 
-from .command_config import CommandConfig, RunnerConfig
+from .command_config import CommandConfig, OutputStorageConfig, RunnerConfig
 from .command_executor import CommandExecutor
 from .command_orchestrator import CommandOrchestrator
 from .command_runtime import CommandRuntime
@@ -34,6 +34,7 @@ __all__ = [
     "ConcurrencyPolicy",
     "load_config",
     "NewRunDecision",
+    "OutputStorageConfig",
     "ResolvedCommand",
     "RunnerConfig",
     "RunHandle",